continuacao

+++ src/main/java/trabalhojava/buscadordemusicasbrasileiras/controller/SongsController.java
@@ -3,6 +3,7 @@ package trabalhojava.buscadordemusicasbrasileiras.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import trabalhojava.buscadordemusicasbrasileiras.exception.SongOperationException;
 import trabalhojava.buscadordemusicasbrasileiras.model.SongEntity;
 import trabalhojava.buscadordemusicasbrasileiras.repository.SongsRepository;
 
@@ -15,10 +16,6 @@ public class SongsController {
     @Autowired
     SongsRepository songsRepository;
 
-    public SongsController(SongsRepository songsRepository) {
-        this.songsRepository = songsRepository;
-    }
-
     @GetMapping("/minhasMusicas")
     public List<SongEntity> getAllSongs() {
         return songsRepository.findAll();
@@ -26,6 +23,9 @@ public class SongsController {
 
     @GetMapping("/minhasMusicas/{id}")
     public ResponseEntity<SongEntity> getSongById(@PathVariable String id) {
+        if(songsRepository.findById(id).isEmpty()) {
+            throw new SongOperationException("Erro: Essa playlist não foi encontrada no perfil.");
+        }
         return songsRepository.findById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -33,11 +33,21 @@ public class SongsController {
 
     @PostMapping("/adicionarMusica")
     public void adicionarMusica(@RequestBody SongEntity song) {
+        if(songsRepository.existsByDeezerId(song.getDeezerId())) {
+            throw new SongOperationException("Erro: Essa playlist já está adicionada no perfil.");
+        }
+        if(songsRepository.count() > 20) {
+            throw new SongOperationException("Erro: A lista de playlists está com capacidade cheia.");
+        }
         songsRepository.save(song);
     }
 
     @DeleteMapping("/deletarMusica/{id}")
     public void deletarMusica(@PathVariable String id) {
+        if(songsRepository.findById(id).isEmpty()) {
+            throw new SongOperationException("Erro: Essa música não foi encontrada no perfil.");
+        }
+
         songsRepository.deleteById(id);
     }
